OC-10474 Oops error page for single user when logged in to multiple studies/sites - Moved the studyEnv code from the MainMenu to SecureController
